+++ digiplan/map/hooks.py
@@ -62,7 +62,7 @@ def adapt_electricity_demand(scenario: str, data: dict, request: HttpRequest) ->
     """
     year = "2045" if scenario == "scenario_2045" else "2022"
     for sector, slider in (("hh", "s_v_2"), ("cts", "s_v_3"), ("ind", "s_v_4")):
-        demand_filename = settings.DATA_DIR.path("scenarios").path(f"demand_{sector}_power_demand.csv")
+        demand_filename = settings.DIGIPIPE_DIR.path("scalars").path(f"demand_{sector}_power_demand.csv")
         demand = pd.read_csv(demand_filename)
         data[f"ABW-electricity-demand_{sector}"] = {"amount": float(demand[year].sum()) * data.pop(slider) / 100}
     return data
@@ -89,10 +89,12 @@ def adapt_heat_demand(scenario: str, data: dict, request: HttpRequest) -> dict: 
     year = "2045" if scenario == "scenario_2045" else "2022"
     for sector, slider in (("hh", "w_v_3"), ("cts", "w_v_4"), ("ind", "w_v_5")):
         percentage = data.pop(slider)
-        for location in ("central", "decentral"):
-            demand_filename = settings.DATA_DIR.path("scenarios").path(f"demand_{sector}_heat_demand_{location}.csv")
+        for dataset_loc, loc in (("cen", "central"), ("dec", "decentral")):
+            demand_filename = settings.DIGIPIPE_DIR.path("scalars").path(
+                f"demand_{sector}_heat_demand_{dataset_loc}.csv",
+            )
             demand = pd.read_csv(demand_filename)
-            data[f"ABW-heat_{location}-demand_{sector}"] = {
+            data[f"ABW-heat_{loc}-demand_{sector}"] = {
                 "amount": float(demand[year].sum()) * percentage / 100,
             }
     return data