+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project tries to adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- datapackage from digipipe
+- models, layers and legend items from digipipe geodata
+
+### Changed
+- paths for oemof hooks to digipipe scalars
+
 ## [0.4.0] - 2023-06-20
 ### Added
 - complete energy settings